# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Local Router API",
    "description": "API for controlling the local router service and forwarding OpenAI-compatible requests",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:11436",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "summary": "Create chat completion",
        "description": "Creates a chat completion request and forwards it to the appropriate provider",
        "tags": [
          "OpenAI Compatible"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "model",
                  "messages"
                ],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Model name in format [provider]model",
                    "example": "[provider1]gpt-3.5-turbo"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "role",
                        "content"
                      ],
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": [
                            "system",
                            "user",
                            "assistant"
                          ]
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "stream": {
                    "type": "boolean",
                    "description": "Whether to stream the response",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "object": {
                      "type": "string",
                      "example": "chat.completion"
                    },
                    "created": {
                      "type": "integer"
                    },
                    "model": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "index": {
                            "type": "integer"
                          },
                          "message": {
                            "type": "object",
                            "properties": {
                              "role": {
                                "type": "string"
                              },
                              "content": {
                                "type": "string"
                              }
                            }
                          },
                          "finish_reason": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid model or request format"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}